+++ src/main/java/com/github/gtexpert/gtwp/integration/ids/IDsConfigHolder.java
@@ -1,0 +1,12 @@
+package com.github.gtexpert.gtwp.integration.ids;
+
+import net.minecraftforge.common.config.Config;
+
+import com.github.gtexpert.gtwp.api.ModValues;
+import com.github.gtexpert.gtwp.module.Modules;
+
+@Config.LangKey(ModValues.MODID + ".config.integration.id")
+@Config(modid = ModValues.MODID,
+        name = ModValues.MODID + "/integration/" + Modules.MODULE_IDS,
+        category = "IntegratedDynamics")
+public class IDsConfigHolder {}